backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[argon2]
python-dotenv
pyotp
authlib
python-multipart
